Fixed issues with custom elements not rendering correctly:
When you would map over an array of elements that contain a custom element, and then sort that array in a different manner, custom elements were not being correctly unmounted and remounted.

 Broke out the tooltip component into two separate components for easier understanding.
