+++ package.json
@@ -68,6 +68,6 @@
     "color": "^1.0.3",
     "history": "1.17.0",
     "moment": "^2.18.1",
-    "preact": "7.2.1"
+    "preact": "^8.2.5"
   }
 }
