+++ src/custom-elements/cps-tooltip/tooltip-popup.styles.css
@@ -12,7 +12,3 @@
 	text-align: center;
 	font-weight: normal;
 }
-
-.inlineBlock {
-	display: inline-block;
-}
